Delete refactored registers code
