--- android/quest/src/main/java/org/smartregister/fhircore/quest/ui/task/PatientTaskListenerIntent.kt
@@ -1,22 +1,0 @@
-/*
- * Copyright 2021 Ona Systems, Inc
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.smartregister.fhircore.quest.ui.task
-
-import org.smartregister.fhircore.engine.util.ListenerIntent
-
-/** [ListenerIntent] used on individual rows click listeners for patient task register. */
-sealed class PatientTaskListenerIntent : ListenerIntent
