+++ android/quest/src/test/java/org/smartregister/fhircore/quest/ui/login/QuestLoginServiceTest.kt
@@ -17,6 +17,7 @@
 package org.smartregister.fhircore.quest.ui.login
 
 import android.content.Intent
+import androidx.compose.material.ExperimentalMaterialApi
 import dagger.hilt.android.testing.HiltAndroidRule
 import dagger.hilt.android.testing.HiltAndroidTest
 import io.mockk.mockk
@@ -36,7 +37,7 @@ import org.smartregister.fhircore.engine.configuration.ConfigurationRegistry
 import org.smartregister.fhircore.engine.ui.login.LoginActivity
 import org.smartregister.fhircore.quest.app.fakes.Faker
 import org.smartregister.fhircore.quest.robolectric.RobolectricTest
-import org.smartregister.fhircore.quest.ui.patient.register.PatientRegisterActivity
+import org.smartregister.fhircore.quest.ui.main.AppMainActivity
 
 @HiltAndroidTest
 @Ignore("Fix appId not initialized")
@@ -67,11 +68,12 @@ class QuestLoginServiceTest : RobolectricTest() {
     loginActivity.finish()
   }
 
+  @OptIn(ExperimentalMaterialApi::class)
   @Test
   fun testNavigateToHomeShouldNavigateToRegisterScreen() {
     loginService.navigateToHome()
     val startedIntent: Intent = shadowOf(loginActivity).nextStartedActivity
     val shadowIntent: ShadowIntent = shadowOf(startedIntent)
-    Assert.assertEquals(PatientRegisterActivity::class.java, shadowIntent.intentClass)
+    Assert.assertEquals(AppMainActivity::class.java, shadowIntent.intentClass)
   }
 }
