+++ android/quest/src/test/java/org/smartregister/fhircore/quest/ui/patient/details/ListDataDetailScreenTest.kt
@@ -63,7 +63,7 @@ import org.smartregister.fhircore.quest.data.patient.model.AdditionalData
 import org.smartregister.fhircore.quest.data.patient.model.PatientItem
 import org.smartregister.fhircore.quest.data.patient.model.QuestResultItem
 import org.smartregister.fhircore.quest.robolectric.RobolectricTest
-import org.smartregister.fhircore.quest.ui.patient.register.PatientItemMapper
+import org.smartregister.fhircore.quest.util.mappers.PatientItemMapper
 
 @HiltAndroidTest
 @Ignore("To be deleted test class; new test to be written after refactor")
