+++ android/quest/src/test/java/org/smartregister/fhircore/quest/ui/patient/details/ListDataDetailViewModelTest.kt
@@ -41,7 +41,7 @@ import org.smartregister.fhircore.quest.app.fakes.Faker
 import org.smartregister.fhircore.quest.data.patient.PatientRepository
 import org.smartregister.fhircore.quest.data.patient.model.QuestResultItem
 import org.smartregister.fhircore.quest.robolectric.RobolectricTest
-import org.smartregister.fhircore.quest.ui.patient.register.PatientItemMapper
+import org.smartregister.fhircore.quest.util.mappers.PatientItemMapper
 
 @HiltAndroidTest
 class ListDataDetailViewModelTest : RobolectricTest() {
