+++ android/quest/src/test/java/org/smartregister/fhircore/quest/ui/patient/details/SimpleDetailsScreenTest.kt
@@ -48,7 +48,7 @@ import org.smartregister.fhircore.quest.app.fakes.Faker
 import org.smartregister.fhircore.quest.data.patient.PatientRepository
 import org.smartregister.fhircore.quest.data.patient.model.DetailsViewItem
 import org.smartregister.fhircore.quest.robolectric.RobolectricTest
-import org.smartregister.fhircore.quest.ui.patient.register.PatientItemMapper
+import org.smartregister.fhircore.quest.util.mappers.PatientItemMapper
 
 @HiltAndroidTest
 class SimpleDetailsScreenTest : RobolectricTest() {
